+++ scripts/get-icons.sh
@@ -23,7 +23,7 @@ fi
 #╔═══════════════════════════════════════════════════════════╗
 #║ Directories                                               ║
 #╚═══════════════════════════════════════════════════════════╝
-icons="${PWD}/public/icons"
+icons="${PWD}/assets/icons"
 common="${icons}/common"
 logos="${icons}/logos"
 mkdir -p "${logos}" "${common}"